Gogeta.SolverParams

Gogeta.TEModel

Gogeta.NN_to_MIP

Gogeta.TE_to_MIP

Gogeta.children

Gogeta.compress

Gogeta.extract_evotrees_info

Gogeta.forward_pass!

Gogeta.get_solution

Gogeta.init_TEModel!

Gogeta.optimize_with_initial_constraints!

Gogeta.optimize_with_lazy_constraints!

Gogeta.SolverParams — Type

SolverParams

Parameters to be used by the solver.

Fields

solver: has to be "Gurobi" or "GLPK"

silent: is the solver log shown

threads: use 0 for solver default

relax: linear relaxation for the MIP

time_limit: time limit for each optimization in the model

Examples

julia> solver_params = SolverParams(solver="Gurobi", silent=true, threads=0, relax=false, time_limit=0);

Gogeta.TEModel — Type

Universal datatype for storing information about a Tree Ensemble Model. This is the datatype that is used when creating the integer optimization problem from a tree ensemble.

Different tree models (EvoTrees, XGBoost, RandomForest) require individual conversion functions to this datatype.

Fields

n_trees: number of trees in the ensemble

n_feats: number of features (input variables) in the model

n_leaves: number of leaves on each tree

leaves: indices of the leaves on each tree

splits: [feature, splitpoint index] pairs accessible by [tree, node]

splits_ordered: splitpoints ordered by split value for each feature

n_splits: number of splitpoints for each feature

predictions: prediction of each node (zero for nodes that are not leaves)

split_nodes: boolean array containing information whether a node is a split node or not

Splitpoints is the set of unique condition values from the ensemble. Each node is associated with a condition value.

Gogeta.NN_to_MIP — Method

function NN_to_MIP(NN_model::Flux.Chain, init_ub::Vector{Float64}, init_lb::Vector{Float64}, solver_params::SolverParams; tighten_bounds::String="fast", bounds_U=nothing, bounds_L=nothing, out_ub=nothing, out_lb=nothing)

Creates a mixed-integer optimization problem from a ReLU-activated neural network.

Returns a JuMP model containing the MIP formulation as well as the upper and lower activation bounds for each neuron.

The MIP can be created with initial bounds (optional arguments), or the bounds can be calculated as the model is created in either "fast" or "standard" mode. If output bounds are to be considered during the tightening, they have to be provided as optional arguments and tighten_bounds must be set to "output".

Arguments

NN_model: neural network as a Flux.Chain

init_ub: upper bounds for the input layer

init_lb: lower bounds for the input layer

solver_params: parameters for the JuMP model solver

Optional arguments

tighten_bounds: "fast", "standard" or "output"

bounds_U: upper bounds for the hidden and output layers

bounds_L: lower bounds for the hidden and output layers

out_ub: upper bounds for the output layer

out_lb: lower bounds for the output layer

Examples

julia> nn_jump, U, L = NN_to_MIP(model, init_U, init_L, solver_params; tighten_bounds="standard");

Gogeta.TE_to_MIP — Method

function TE_to_MIP(TE::TEModel, optimizer, objective)

Creates a JuMP model opt_model based on the given tree ensemble. Returns opt_model.

The JuMP model is created without the split constraints.

Arguments

TE: A tree ensemble model in the universal data type TEModel.

optimizer: Optimizer object that will be given to the JuMP model.

objective: MINSENSE or MAXSENSE.

Gogeta.children — Method

function children(id::Int, leaf_dict::Dict, max::Int)

Finds the leaf indices of the children leaves of node id in a binary tree.

Returns an array of the leaf indices.

Arguments

id: Index of the node in a binary tree. Indexing starts from one and follows level order.

leaf_dict: A dictionary (map) of the leaf indices accessible by the node indices.

max: Biggest possible node id in the tree. Used to terminate the search.

Gogeta.compress — Method

function compress(model::Flux.Chain, init_ub::Vector{Float64}, init_lb::Vector{Float64}; params=nothing, bounds_U=nothing, bounds_L=nothing, tighten_bounds="fast")

Creates a new neural network model by identifying stabily active and inactive neurons and removing them.

Can be called with precomputed bounds. Returns the compressed neural network as a Flux.Chain and the indices of the removed neurons in this case.

Can also be called without the bounds to invoke bound tightening ("standard" or "fast" mode). In this case solver parameters have to be provided. Returns the resulting JuMP model, the compressed neural network, the removed neurons and the computed bounds.

Arguments

NN_model: neural network as a Flux.Chain

init_ub: upper bounds for the input layer

init_lb: lower bounds for the input layer

Optional arguments

params: parameters for the JuMP model solver

tighten_bounds: "fast" or "standard"

bounds_U: upper bounds for the hidden and output layers

bounds_L: lower bounds for the hidden and output layers

Examples

julia> jump_model, compressed_model, removed_neurons, bounds_U, bounds_L = compress(model, init_U, init_L; params=solver_params, tighten_bounds="standard");

Gogeta.extract_evotrees_info — Method

extract_evotrees_info(evo_model; tree_limit=length(evo_model.trees))

Gets the data required for constructing the corresponding MIP from an EvoTrees model evo_model. Returns a custom datatype TEModel which contains the necessary information.

Arguments

evo_model: A trained EvoTrees tree ensemble model.

Optional arguments

tree_limit: only first n trees specified by the argument will be used

Gogeta.forward_pass! — Method

function forward_pass!(jump_model::JuMP.Model, input)

Calculates the output of a neural network -representing JuMP model given some input.

Gogeta.get_solution — Method

function get_solution(model::JuMP.Model, TE::TEModel)

Finds the upper and lower bounds for each input variable given the optimized model.

Returns the bounds for each feature in an array.

Arguments

model: The optimized JuMP model.

TE: Struct of type TEModel containing information about the tree ensemble.

Gogeta.init_TEModel! — Method

function init_TEModel!(TE::TEModel)

Precompute child leaves which are needed for generating the split constraints. Changes child_leaves field of the TEModel.

Arguments

TE: Struct of type TEModel containing information about the tree ensemble.

Gogeta.optimize_with_initial_constraints! — Method

function optimize_with_initial_constraints!(opt_model::JuMP.Model, TE::TEModel)

Adds all split constraints to the formulation and then solves the MIP.

Arguments

opt_model: A JuMP model containing the formulation.

TE: A tree ensemble model in the universal data type TEModel.

Gogeta.optimize_with_lazy_constraints! — Method

function optimize_with_lazy_constraints!(opt_model::JuMP.Model, TE::TEModel)

Solves the optimization model by utilizing lazy constraints. This means that the split constraints are added one-by-one for each tree.

Arguments

opt_model: A JuMP model containing the formulation.

TE: A tree ensemble model in the universal data type TEModel.