Gogeta.SolverParams

Gogeta.TEModel

Gogeta.NN_to_MIP_with_bound_tightening

Gogeta.NN_to_MIP_with_precomputed

Gogeta.TE_to_MIP

Gogeta.build_model!

Gogeta.calculate_bounds

Gogeta.children

Gogeta.compress_with_bound_tightening

Gogeta.compress_with_precomputed

Gogeta.copy_model

Gogeta.extract_evotrees_info

Gogeta.formulate_and_compress

Gogeta.forward_pass!

Gogeta.get_solution

Gogeta.init_TEModel!

Gogeta.optimize_with_initial_constraints!

Gogeta.optimize_with_lazy_constraints!

Gogeta.prune!

Gogeta.set_solver_params!

Gogeta.SolverParams — Type

SolverParams

Parameters to be used by the solver.

Fields

solver: has to be "Gurobi" or "GLPK"

silent: is the solver log shown

threads: use 0 for solver default

relax: linear relaxation for the MIP

time_limit: time limit for each optimization in the model

Examples

julia> solver_params = SolverParams(solver="Gurobi", silent=true, threads=0, relax=false, time_limit=0);

Gogeta.TEModel — Type

struct TEModel

Universal datatype for storing information about a Tree Ensemble Model. This is the datatype that is used when creating the integer optimization problem from a tree ensemble.

Different tree models (EvoTrees, XGBoost, RandomForest) require individual conversion functions to this datatype.

Fields

n_trees: number of trees in the ensemble

n_feats: number of features (input variables) in the model

n_leaves: number of leaves on each tree

leaves: indices of the leaves on each tree

splits: [feature, splitpoint index] pairs accessible by [tree, node]

splits_ordered: splitpoints ordered by split value for each feature

n_splits: number of splitpoints for each feature

predictions: prediction of each node (zero for nodes that are not leaves)

split_nodes: boolean array containing information whether a node is a split node or not

Splitpoints is the set of unique condition values from the ensemble. Each node is associated with a condition value.

Gogeta.NN_to_MIP_with_bound_tightening — Method

function NN_to_MIP_with_bound_tightening(NN_model::Flux.Chain, U_in, L_in, solver_params; bound_tightening="fast", U_out=nothing, L_out=nothing, silent=false)

Creates a mixed-integer optimization problem from a Flux.Chain model. In this version, the neuron activation bounds are calculated as the model is being created.

Returns the resulting JuMP model and the computed activation bounds.

Parameters

NN_model: the neural network as a Flux.Chain

U_in: upper bounds of the input neurons

L_in: lower bounds of the input neurons

solver_params: a SolverParams struct

Optional arguments

bound_tightening: "fast", "standard" or "output"

silent: controls the output logs

U_out: upper activation bounds for the output layer (must be set if bound_tightening="output")

L_out: lower activation bounds for the output layer (must be set if bound_tightening="output")

Gogeta.NN_to_MIP_with_precomputed — Method

function NN_to_MIP_with_precomputed(NN_model::Flux.Chain, U_in, L_in, solver_params, U_bounds, L_bounds; silent=false)

Creates a mixed-integer optimization problem from a Flux.Chain model. In this version, the neuron activation bounds must be given as arguments.

Returns the resulting JuMP model.

Parameters

NN_model: the neural network as a Flux.Chain

U_in: upper bounds of the input neurons

L_in: lower bounds of the input neurons

solver_params: a SolverParams struct

U_bounds: precomputed upper activation bounds for each layer

L_bounds: precomputed lower activation bounds for each layer

Optional arguments

silent: controls the output logs

Gogeta.TE_to_MIP — Method

function TE_to_MIP(TE::TEModel, optimizer, objective)

Creates a JuMP model opt_model based on the given tree ensemble. Returns opt_model.

The JuMP model is created without the split constraints.

Arguments

TE: A tree ensemble model in the universal data type TEModel.

optimizer: Optimizer object that will be given to the JuMP model.

objective: MINSENSE or MAXSENSE.

Gogeta.build_model! — Method

function build_model!(W, b, K, neurons)

Builds a new Flux.Chain model from the given weights and biases.

Gogeta.calculate_bounds — Method

function calculate_bounds(model::JuMP.Model, layer, neuron, W, b, neurons; layers_removed=0)

Calculates the upper and lower activation bounds for a neuron in a ReLU-activated neural network.

Gogeta.children — Method

function children(id::Int, leaf_dict::Dict, max::Int)

Finds the leaf indices of the children leaves of node id in a binary tree.

Returns an array of the leaf indices.

Arguments

id: Index of the node in a binary tree. Indexing starts from one and follows level order.

leaf_dict: A dictionary (map) of the leaf indices accessible by the node indices.

max: Biggest possible node id in the tree. Used to terminate the search.

Gogeta.compress_with_bound_tightening — Method

function compress_with_bound_tightening(NN_model::Flux.Chain, U_in, L_in, solver_params; bound_tightening="fast", silent=false)

Creates a compressed version of the neural network model Flux.Chain given as input. This is accomplished by identifying inactive and stabily active neurons. This version computes the bounds as the network is being compressed.

Return the compressed model, the list of removed neurons, the JuMP model and computed bounds.

Parameters

NN_model: the neural network as a Flux.Chain

U_in: upper bounds of the input neurons

L_in: lower bounds of the input neurons

solver_params: a SolverParams struct

Optional arguments

bound_tightening: "fast" or "standard"

silent: controls the output logs

Gogeta.compress_with_precomputed — Method

function compress_with_precomputed(NN_model::Flux.Chain, U_in, L_in, U_bounds, L_bounds; silent=true)

Creates a compressed version of the neural network model Flux.Chain given as input. This is accomplished by identifying inactive and stabily active neurons.

Return the compressed model and the list of removed neurons.

Parameters

NN_model: the neural network as a Flux.Chain

U_in: upper bounds of the input neurons

L_in: lower bounds of the input neurons

U_bounds: precomputed upper activation bounds for each layer

L_bounds: precomputed lower activation bounds for each layer

Optional arguments

silent: controls the output logs

Gogeta.copy_model — Method

function copy_model(input_model, solver_params)

Copies a JuMP model. Solver has to be specified for each new copy. Used for parallelization.

Gogeta.extract_evotrees_info — Method

extract_evotrees_info(evo_model; tree_limit=length(evo_model.trees))

Gets the data required for constructing the corresponding MIP from an EvoTrees model evo_model. Returns a custom datatype TEModel which contains the necessary information.

Arguments

evo_model: A trained EvoTrees tree ensemble model.

Optional arguments

tree_limit: only first n trees specified by the argument will be used

Gogeta.formulate_and_compress — Method

function formulate_and_compress(NN_model::Flux.Chain, U_in, L_in; U_bounds=nothing, L_bounds=nothing, U_out=nothing, L_out=nothing, solver_params=nothing, bound_tightening="fast", compress=false, silent=false)

Creates a mixed-integer optimization problem from a Flux.Chain model.

The parameters are used to specify what kind of bound tightening and compression will be used.

Gogeta.forward_pass! — Method

function forward_pass!(jump_model::JuMP.Model, input)

Calculates the output of a neural network -representing JuMP model given some input.

Gogeta.get_solution — Method

function get_solution(model::JuMP.Model, TE::TEModel)

Finds the upper and lower bounds for each input variable given the optimized model.

Returns the bounds for each feature in an array.

Arguments

model: The optimized JuMP model.

TE: Struct of type TEModel containing information about the tree ensemble.

Gogeta.init_TEModel! — Method

function init_TEModel!(TE::TEModel)

Precompute child leaves which are needed for generating the split constraints. Changes child_leaves field of the TEModel.

Gogeta.optimize_with_initial_constraints! — Method

function optimize_with_initial_constraints!(opt_model::JuMP.Model, TE::TEModel)

Adds all split constraints to the formulation and then solves the MIP.

Arguments

opt_model: A JuMP model containing the formulation.

TE: A tree ensemble model in the universal data type TEModel.

Gogeta.optimize_with_lazy_constraints! — Method

function optimize_with_lazy_constraints!(opt_model::JuMP.Model, TE::TEModel)

Solves the optimization model by utilizing lazy constraints. This means that the split constraints are added one-by-one for each tree.

Arguments

opt_model: A JuMP model containing the formulation.

TE: A tree ensemble model in the universal data type TEModel.

Gogeta.prune! — Method

function prune!(W, b, removed_neurons, layers_removed, neuron_count, layer, bounds_U, bounds_L)

Removes stabily active or inactive neurons in a network by updating the weights and the biases and the removed neurons list accordingly.

Gogeta.set_solver_params! — Method

function set_solver_params!(model, params)

Set the parameters of a JuMP model. Solver and its parameters have to be specified for each new model copy. Used for parallelization.