Gogeta.TEModel

Gogeta.bound_tightening

Gogeta.bound_tightening_2workers

Gogeta.bound_tightening_threads

Gogeta.bound_tightening_workers

Gogeta.children

Gogeta.create_CNN_JuMP_model

Gogeta.create_JuMP_model

Gogeta.evaluate!

Gogeta.evaluate_CNN!

Gogeta.extract_evotrees_info

Gogeta.get_solution

Gogeta.tree_model_to_MIP

Gogeta.TEModel — Type

Universal datatype for storing information about a Tree Ensemble Model.

Fields

n_trees - number of trees in the ensemble

n_feats - number of features (variables) in the model

n_leaves - number of leaves on each tree

leaves - ids of the leaves on each tree

splits - [feature, splitpoint index] pairs accessible by [tree, node]

splits_ordered - splitpoints ordered by split value for each feature

n_splits - number of splitpoints for each feature

predictions - prediction of each node (zero for nodes that are not leaves)

split_nodes - boolean array containing information whether a node is a split node or not

Gogeta.bound_tightening — Function

boundtightening(DNN::Chain, initUbounds::Vector{Float32}, initL_bounds::Vector{Float32}, verbose::Bool=false)

A single-threaded implementation of optimal tightened constraint bounds L and U for for a trained DNN. Using these bounds with the createJuMPmodel function reduces solution time for optimization problems.

Arguments

DNN::Chain: A trained ReLU DNN.

init_U_bounds::Vector{Float32}: Initial upper bounds on the node values of the DNN.

init_L_bounds::Vector{Float32}: Initial lower bounds on the node values of the DNN.

verbose::Bool=false: Controls Gurobi logs.

tl::Float64=1.0: Controls the time limit for solvign the subproblems

Examples

L_bounds, U_bounds = bound_tightening(DNN, init_U_bounds, init_L_bounds, false, 1.0)

Gogeta.bound_tightening_2workers — Function

boundtightening2workers(DNN::Chain, initUbounds::Vector{Float32}, initLbounds::Vector{Float32}, verbose::Bool=false)

A multi-threaded (using workers) implementation of optimal tightened constraint bounds L and U for for a trained DNN. This function uses two in-place models at each layer to reduce memory usage. A max of 2 workers in use simultaneously. Using these bounds with the createJuMPmodel function reduces solution time for optimization problems.

Arguments

DNN::Chain: A trained ReLU DNN.

init_U_bounds::Vector{Float32}: Initial upper bounds on the node values of the DNN.

init_L_bounds::Vector{Float32}: Initial lower bounds on the node values of the DNN.

verbose::Bool=false: Controls Gurobi logs.

tl::Float64=1.0: Controls the time limit for solvign the subproblems.

Examples

L_bounds_workers, U_bounds_workers = bound_tightening_2workers(DNN, init_U_bounds, init_L_bounds, false, 1.0)

Gogeta.bound_tightening_threads — Function

boundtighteningthreads(DNN::Chain, initUbounds::Vector{Float32}, initLbounds::Vector{Float32}, verbose::Bool=false, tl::float64=1)

A multi-threaded (using Threads) implementation of optimal tightened constraint bounds L and U for for a trained DNN. Using these bounds with the createJuMPmodel function reduces solution time for optimization problems.

Arguments

DNN::Chain: A trained ReLU DNN.

init_U_bounds::Vector{Float32}: Initial upper bounds on the node values of the DNN.

init_L_bounds::Vector{Float32}: Initial lower bounds on the node values of the DNN.

verbose::Bool=false: Controls Gurobi logs.

tl::Float64=1.0: Controls the time limit for solvign the subproblems

Examples

L_bounds_threads, U_bounds_threads = bound_tightening_threads(DNN, init_U_bounds, init_L_bounds, false, 1.0)

Gogeta.bound_tightening_workers — Function

boundtighteningworkers(DNN::Chain, initUbounds::Vector{Float32}, initLbounds::Vector{Float32}, verbose::Bool=false, tl::Float64=1.0)

A multi-threaded (using workers) implementation of optimal tightened constraint bounds L and U for for a trained DNN. Using these bounds with the createJuMPmodel function reduces solution time for optimization problems.

Arguments

DNN::Chain: A trained ReLU DNN.

init_U_bounds::Vector{Float32}: Initial upper bounds on the node values of the DNN.

init_L_bounds::Vector{Float32}: Initial lower bounds on the node values of the DNN.

verbose::Bool=false: Controls Gurobi logs.

tl::Float64=1.0: Controls the time limit for solvign the subproblems

Examples

L_bounds_workers, U_bounds_workers = bound_tightening_workers(DNN, init_U_bounds, init_L_bounds, false, 1.0)

Gogeta.children — Method

Finds the children leaves of node id in a binary tree.

Gogeta.create_CNN_JuMP_model — Method

createCNNJuMPmodel(CNN::Chain, datashape::Tuple{Int64, Int64, Int64, Int64}, Lbounds::Vector{Array{Float32}}, Ubounds::Vector{Array{Float32}})

Converts a CNN with ReLU activation functions to a 0-1 MILP JuMP model. The ReLU CNN is assumed to be a Flux.Chain. The activation function must be "relu" in all hidden (Conv and Dense) layers and "identity" in the output layer. The model assumes the following things from the underlying CNN:

Input must be Array{Float32, 4}, (e.g. {32, 32, 3, 1}) where first two indices are height and width of data,

third index is channel count (e.g. 1 for grayscale image, 3 for RGB), fourth index is batch size (here must be 1)

The CNN must consist of 3 sections: the convolutional and pooling layers, Flux.flatten, and dense layers.

Convolutional layers must use relu, pooling layers must be MaxPool or MeanPool.

Layers must use default setting, such as stride, pad, dilation, etc. Conv.filter, MaxPool.k and MeanPool.k (window) sizes can be arbitrary.

No convolutional or pooling layers are necessary before Flux.flatten. Also, no dense layers are necessary after Flux.flatten.

Arguments

CNN::Chain: A trained ReLU CNN with the above assumptions

data_shape::Tuple{Int64, Int64, Int64, Int64}: Shape of the data used in the CNN as a Tuple, e.g., (32, 32, 3, 1) (similar logis as above)

L_bounds::Vector{Array{Float32}}: Lower bound big-M values for contraint bounds, indexed L_bounds[layer][channel, height, width]

U_bounds::Vector{Array{Float32}}: Uower bound big-M values for contraint bounds, indexed U_bounds[layer][channel, height, width]

Examples

model = create_CNN_JuMP_model(CNN, data_shape, L_bounds, U_bounds)

Gogeta.create_JuMP_model — Function

create_JuMP_model(DNN::Chain, L_bounds::Vector{Float32}, U_bounds::Vector{Float32}, bound_tightening::String="none", bt_verbose::Bool=false)

Converts a ReLU DNN to a 0-1 MILP JuMP model. The ReLU DNN is assumed to be a Flux.Chain. The activation function must be "relu" in all hidden layers and "identity" in the output layer. The lower and upper bounds to the function are given as a Vector{Float32}. The bounds are in order from the input layer to the output layer. The keyword argument "bt" determines if bound tightening is to be used on the constraint bounds: — "none": No bound tightening is used. — "singletread": One shared JuMP model is used to calculate bounds one at a time. — "threads": The bounds are calculated using a separate model for each subproblem using Threads — "workers": The bounds are calculated using a separate model for each subproblem using Workers — "2 workers": The bounds are calculated with a maximum of two workers at each layer (upper and lower bounds). Each worker reuses the JuMP model.

Arguments

DNN::Chain: A trained ReLU DNN.

L_bounds::Vector{Float32}: Lower bounds on the node values of the DNN.

U_bounds::Vector{Float32}: Upper bounds on the node values of the DNN.

bt::String="none": Optional bound tightening of the constraint bounds. Can be set to "none", "singlethread", "threads", "workers" or "2 workers".

bt_verbose::Bool=false: Controls Gurobi logs in bound tightening procedures.

Examples

model = create_JuMP_model(DNN, L_bounds, U_bounds, "singlethread", false)

Gogeta.evaluate! — Method

evaluate!(JuMP_model::Model, input::Vector{Float32})

Fixes the variables corresponding to the DNN input to a given input vector.

Arguments

JuMP_model::Model: A JuMP model representing a traied ReLU DNN (generated using the function createJuMPmodel).

input::Vector{Float32}: A given input to the trained DNN.

Examples

evaluate!(JuMP_model, input)

Gogeta.evaluate_CNN! — Method

evaluateCNN!(CNNmodel::Model, input::Array{Float32, 4})

Fixes the variables corresponding to the CNN input to a given input array.

Arguments

CNN_model::Model: A JuMP model representing a traied ReLU DNN (generated using the function createJuMPmodel).

input::Array{Float32, 4}: A given input array to the trained CNN.

Examples

evaluate_CNN!(CNN_model, input)

Gogeta.extract_evotrees_info — Method

Gets the data required for constructing the corresponding MIP from an EvoTrees model evo_model. Returns a custom datatype TEModel which contains the necessary information.

Gogeta.get_solution — Method

Gets human-readable array solution where upper and lower bounds for each feature are given.

Gogeta.tree_model_to_MIP — Method

Creates a JuMP model opt_model and optimizes it when given a tree_model in the universal data type TEModel. Returns opt_model, and an array given by get_solution or nothing in the case of a failed solve.

Optional arguments

create_initial - controls whether split constraint algorithm or initial generation is used

objective - MIN_SENSE or default MAX_SENSE

gurobi_env - environment for Gurobi solver

show_output - controls whether output of Gurobi is shown

timelimit - time after which Gurobi solver terminates